python/dataengineeringai/fourth/app/requirements.txt
Flask
minio
requests
psycopg2-binary
cryptography
